Refactor out common summary functions into mixin
