+++ frontend/src/views/c-summary.vue
@@ -64,13 +64,12 @@
 
 <script lang='ts'>
 import { mapState } from 'vuex';
-import { PropType, defineComponent } from 'vue';
+import { defineComponent } from 'vue';
 
 import cErrorMessageBox from '../components/c-error-message-box.vue';
 import cSummaryCharts from '../components/c-summary-charts.vue';
 import cFileTypeCheckboxes from '../components/c-file-type-checkboxes.vue';
 import cSummaryHeader from '../components/c-summary-header.vue';
-import getNonRepeatingColor from '../utils/random-color-generator';
 import sortFiltered from '../utils/repo-sorter';
 import {
   Commit,
@@ -78,49 +77,32 @@ import {
   Repo,
   User,
   isCommit,
-  CommitResult,
 } from '../types/types';
-import { ErrorMessage } from '../types/zod/summary-type';
 import {
   AuthorDailyContributions,
   AuthorFileTypeContributions,
-  FileTypeAndContribution,
 } from '../types/zod/commits-type';
 import { ZoomInfo } from '../types/vuex.d';
 import {
   FilterGroupSelection, FilterTimeFrame, SortGroupSelection, SortWithinGroupSelection,
 } from '../types/summary';
+import summaryMixin from "../mixin/summaryMixin";
 
 const dateFormatRegex = /([12]\d{3}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01]))$/;
 
 export default defineComponent({
   name: 'c-summary',
+
   components: {
     cErrorMessageBox,
     cSummaryCharts,
     cFileTypeCheckboxes,
     cSummaryHeader,
   },
-  props: {
-    repos: {
-      type: Array<Repo>,
-      required: true,
-    },
-    errorMessages: {
-      type: Object as PropType<ErrorMessage>,
-      default() {
-        return {};
-      },
-    },
-    isWidgetMode: {
-      type: Boolean,
-      default: false,
-    },
-  },
+
+  mixins: [summaryMixin],
+
   data(): {
-    checkedFileTypes: Array<string>,
-    fileTypes: Array<string>,
-    filtered: Array<Array<User>>,
     filterSearch: string,
     filterGroupSelection: FilterGroupSelection,
     sortGroupSelection: SortGroupSelection,
@@ -130,7 +112,6 @@ export default defineComponent({
     sortingWithinOption: string,
     isSortingWithinDsc: boolean,
     filterTimeFrame: FilterTimeFrame,
-    filterBreakdown: boolean,
     tmpFilterSinceDate: string,
     tmpFilterUntilDate: string,
     hasModifiedSinceDate: boolean,
@@ -138,18 +119,11 @@ export default defineComponent({
     filterHash: string,
     minDate: string,
     maxDate: string,
-    fileTypeColors: { [key: string]: string },
-    isSafariBrowser: boolean,
-    filterGroupSelectionWatcherFlag: boolean,
     chartGroupIndex: number | undefined,
     chartIndex: number | undefined,
     viewRepoTags: boolean,
-    optimiseTimeline: boolean,
   } {
     return {
-      checkedFileTypes: [] as Array<string>,
-      fileTypes: [] as Array<string>,
-      filtered: [] as Array<Array<User>>,
       filterSearch: '',
       filterGroupSelection: FilterGroupSelection.GroupByRepos,
       sortGroupSelection: SortGroupSelection.GroupTitleDsc, // UI for sorting groups
@@ -159,7 +133,6 @@ export default defineComponent({
       sortingWithinOption: '',
       isSortingWithinDsc: false,
       filterTimeFrame: FilterTimeFrame.Commit,
-      filterBreakdown: window.isPortfolio, // Auto select filter breakdown if portfolio
       tmpFilterSinceDate: '',
       tmpFilterUntilDate: '',
       hasModifiedSinceDate: window.isSinceDateProvided,
@@ -167,37 +140,13 @@ export default defineComponent({
       filterHash: '',
       minDate: window.sinceDate,
       maxDate: window.untilDate,
-      fileTypeColors: {} as { [key: string]: string },
-      isSafariBrowser: /.*Version.*Safari.*/.test(navigator.userAgent),
-      filterGroupSelectionWatcherFlag: false,
       chartGroupIndex: undefined as number | undefined,
       chartIndex: undefined as number | undefined,
       viewRepoTags: false,
-      optimiseTimeline: window.isPortfolio, // Auto select trim timeline if portfolio
     };
   },
+
   computed: {
-    avgContributionSize(): number {
-      let totalLines = 0;
-      let totalCount = 0;
-      this.repos.forEach((repo) => {
-        repo.users?.forEach((user) => {
-          if (user.checkedFileTypeContribution === undefined || user.checkedFileTypeContribution === 0) {
-            this.updateCheckedFileTypeContribution(user);
-          }
-
-          if (user.checkedFileTypeContribution && user.checkedFileTypeContribution > 0) {
-            totalCount += 1;
-            totalLines += user.checkedFileTypeContribution;
-          }
-        });
-      });
-      if (totalCount === 0) {
-        return 0;
-      }
-      return totalLines / totalCount;
-    },
-
     allGroupsMerged: {
       get(): boolean {
         if (this.mergedGroups.length === 0) {
@@ -236,8 +185,8 @@ export default defineComponent({
 
     ...mapState(['mergedGroups']),
   },
-  watch: {
 
+  watch: {
     filterGroupSelection(): void {
       // Deactivates watcher
       if (!this.filterGroupSelectionWatcherFlag) {
@@ -269,6 +218,7 @@ export default defineComponent({
       },
     },
   },
+
   created(): void {
     this.processFileTypes();
     this.renderFilterHash();
@@ -278,25 +228,8 @@ export default defineComponent({
       this.restoreZoomFiltered(zoomInfo);
     }
   },
-  mounted(): void {
-    // Delay execution of filterGroupSelection watcher
-    // to prevent clearing of merged groups
-    setTimeout(() => {
-      this.filterGroupSelectionWatcherFlag = true;
-    }, 0);
-  },
+
   methods: {
-    // view functions //
-    getRepoSenseHomeLink(): string {
-      const version = window.repoSenseVersion;
-      if (!version) {
-        return `${window.HOME_PAGE_URL}/RepoSense/`;
-      }
-      return `${window.HOME_PAGE_URL}`;
-    },
-    getLogoPath(): string {
-      return window.LOGO_PATH;
-    },
     // model functions //
     setSummaryHash(): void {
       const { addHash, encodeHash, removeHash } = window;
@@ -611,36 +544,6 @@ export default defineComponent({
       });
     },
 
-    processFileTypes(): void {
-      const selectedColors = ['#ffe119', '#4363d8', '#3cb44b', '#f58231', '#911eb4', '#46f0f0', '#f032e6',
-        '#bcf60c', '#fabebe', '#008080', '#e6beff', '#9a6324', '#fffac8', '#800000', '#aaffc3', '#808000', '#ffd8b1',
-        '#000075', '#808080'];
-      const fileTypeColors = {} as { [key: string]: string };
-      let i = 0;
-
-      this.repos.forEach((repo) => {
-        repo.users?.forEach((user) => {
-          Object.keys(user.fileTypeContribution).forEach((fileType) => {
-            if (!Object.prototype.hasOwnProperty.call(fileTypeColors, fileType)) {
-              if (i < selectedColors.length) {
-                fileTypeColors[fileType] = selectedColors[i];
-                i += 1;
-              } else {
-                fileTypeColors[fileType] = getNonRepeatingColor(Object.values(fileTypeColors));
-              }
-            }
-            if (!this.fileTypes.includes(fileType)) {
-              this.fileTypes.push(fileType);
-            }
-          });
-        });
-        this.fileTypeColors = fileTypeColors;
-      });
-
-      this.checkedFileTypes = this.fileTypes.slice();
-      this.$store.commit('updateFileTypeColors', this.fileTypeColors);
-    },
-
     splitCommitsWeek(user: User, sinceDate: string, untilDate: string): void {
       const { commits } = user;
       if (commits === undefined) {
@@ -705,92 +608,6 @@ export default defineComponent({
         commit.commitResults.forEach((commitResult) => week.commitResults.push(commitResult));
       }
     },
-
-    getUserCommits(user: User, sinceDate: string, untilDate: string): null {
-      user.commits = [];
-      const userFirst = user.dailyCommits[0];
-      const userLast = user.dailyCommits[user.dailyCommits.length - 1];
-
-      if (!userFirst) {
-        return null;
-      }
-
-      if (!sinceDate || sinceDate === 'undefined') {
-        sinceDate = userFirst.date;
-      }
-
-      if (!untilDate) {
-        untilDate = userLast.date;
-      }
-
-      user.dailyCommits.forEach((commit) => {
-        const { date } = commit;
-        if (date >= sinceDate && date <= untilDate) {
-          const filteredCommit: DailyCommit = JSON.parse(JSON.stringify(commit));
-          this.filterCommitByCheckedFileTypes(filteredCommit);
-
-          if (filteredCommit.commitResults.length > 0) {
-            filteredCommit.commitResults.forEach((commitResult) => {
-              if (commitResult.messageBody !== '') {
-                commitResult.isOpen = true;
-              }
-            });
-            // The typecast is safe here as we add the insertions and deletions fields
-            // in the filterCommitByCheckedFileTypes method above
-            user.commits?.push(filteredCommit as Commit);
-          }
-        }
-      });
-
-      return null;
-    },
-
-    filterCommitByCheckedFileTypes(commit: DailyCommit): void {
-      let commitResults = commit.commitResults.map((result) => {
-        const filteredFileTypes = this.getFilteredFileTypes(result);
-        this.updateCommitResultWithFileTypes(result, filteredFileTypes);
-        return result;
-      });
-
-      if (!this.isAllFileTypesChecked()) {
-        commitResults = commitResults.filter(
-          (result) => Object.values(result.fileTypesAndContributionMap).length > 0,
-        );
-      }
-
-      // Typecast from DailyCommit to Commit as we add insertions and deletions fields
-      (commit as Commit).insertions = commitResults.reduce((acc, result) => acc + result.insertions, 0);
-      (commit as Commit).deletions = commitResults.reduce((acc, result) => acc + result.deletions, 0);
-      commit.commitResults = commitResults;
-    },
-
-    getFilteredFileTypes(commitResult: CommitResult): { [key: string]: { insertions: number; deletions: number; }; } {
-      return Object.keys(commitResult.fileTypesAndContributionMap)
-        .filter(this.isFileTypeChecked)
-        .reduce((obj: { [key: string]: FileTypeAndContribution }, fileType) => {
-          obj[fileType] = commitResult.fileTypesAndContributionMap[fileType];
-          return obj;
-        }, {});
-    },
-
-    isFileTypeChecked(fileType: string): boolean {
-      if (this.filterBreakdown) {
-        return this.checkedFileTypes.includes(fileType);
-      }
-      return true;
-    },
-
-    updateCommitResultWithFileTypes(
-      commitResult: CommitResult,
-      filteredFileTypes: { [key: string]: FileTypeAndContribution },
-    ): void {
-      commitResult.insertions = Object.values(filteredFileTypes)
-        .reduce((acc, fileType) => acc + fileType.insertions, 0);
-      commitResult.deletions = Object.values(filteredFileTypes)
-        .reduce((acc, fileType) => acc + fileType.deletions, 0);
-      commitResult.fileTypesAndContributionMap = filteredFileTypes;
-    },
-
     getOptionWithOrder(): void {
       const [sortingOption, isSortingDsc] = this.sortGroupSelection.split(' ');
       this.sortingOption = sortingOption;
@@ -810,18 +627,6 @@ export default defineComponent({
       window.removeHash('since');
       window.removeHash('until');
       this.getFiltered();
-    },
-
-    updateCheckedFileTypeContribution(ele: User): void {
-      let validCommits = 0;
-      Object.keys(ele.fileTypeContribution).forEach((fileType) => {
-        if (!this.filterBreakdown) {
-          validCommits += ele.fileTypeContribution[fileType];
-        } else if (this.checkedFileTypes.includes(fileType)) {
-          validCommits += ele.fileTypeContribution[fileType];
-        }
-      });
-      ele.checkedFileTypeContribution = validCommits;
     },
 
     restoreZoomFiltered(info: ZoomInfo): void {
@@ -861,17 +666,6 @@ export default defineComponent({
       info.isRefreshing = false;
       this.$store.commit('updateTabZoomInfo', info);
     },
-    matchZoomUser(info: ZoomInfo, user: User): boolean {
-      const {
-        zIsMerged, zFilterGroup, zRepo, zAuthor,
-      } = info;
-      if (zIsMerged) {
-        return zFilterGroup === 'groupByRepos'
-          ? user.repoName === zRepo
-          : user.name === zAuthor;
-      }
-      return user.repoName === zRepo && user.name === zAuthor;
-    },
 
     dateRounding(datestr: string, roundDown: number): string {
       // rounding up to nearest monday
@@ -885,10 +679,6 @@ export default defineComponent({
       }
 
       return window.getDateStr(datems);
-    },
-
-    isAllFileTypesChecked(): boolean {
-      return this.checkedFileTypes.length === this.fileTypes.length;
     },
   },
 });
